requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
chromadb==0.5.5
streamlit==1.38.0
ollama==0.3.3
python-multipart==0.0.9
pypdf==4.3.1
python-docx==1.1.2
sentence-transformers==3.1.0
python-dotenv==1.0.1
httpx==0.27.2
